Added a11y-utils package and added partial code documentation for focus trap
