+++ tools/a11y-utils/.babelrc
@@ -1,0 +1,3 @@
+{
+	"presets": ["babel-preset-rhythm-ui"]
+}
