+++ tools/a11y-utils/package.json
@@ -1,0 +1,23 @@
+{
+	"name": "@rhythm-ui/a11y-utils",
+	"version": "1.0.0",
+	"main": "lib/index.js",
+	"license": "BSD-3-Clause",
+	"scripts": {
+		"start": "yarn clean && concurrently \"yarn start:babel\" \"yarn start:declarations\"",
+		"start:babel": "babel -w -s -d lib src --extensions \".ts,.tsx\" --ignore *.spec.ts --copy-files",
+		"start:declarations": "tsc --declaration --emitDeclarationOnly --watch",
+		"clean": "rimraf lib",
+		"build": "yarn clean && yarn build:babel && yarn build:declarations",
+		"build:babel": "babel -s -d lib src --extensions \".ts,.tsx\" --ignore *.spec.ts --ignore *.d.ts --copy-files",
+		"build:declarations": "tsc --declaration --emitDeclarationOnly"
+	},
+	"files": [
+		"lib/"
+	],
+	"devDependencies": {
+		"babel-preset-rhythm-ui": "^1.0.0"
+	},
+	"dependencies": {
+	}
+}
