+++ tools/a11y-utils/src/index.ts
@@ -1,0 +1,8 @@
+import FocusTrap from './FocusTrap';
+
+
+export default {
+    FocusTrap
+}
+
+export {FocusTrap};
