+++ tools/a11y-utils/tsconfig.json
@@ -1,0 +1,10 @@
+{
+	"extends": "../../tsconfig.json",
+	"compilerOptions": {
+		"outDir": "./lib",
+		"declarationDir": "./lib"
+	},
+	"include": [
+		"./src"
+	]
+}
